+++ harrypotter/src/main/java/com/isep/harrypotter/model/characters/Enemy.java
@@ -1,6 +1,4 @@
-package com.isep.harrypotter.characters;
-
-import com.isep.harrypotter.characters.AbstractEnemy;
+package com.isep.harrypotter.model.characters;
 
 public class Enemy extends AbstractEnemy {
     public Enemy(int totalHealth, double currentHealth, int defence, int damage , double accuracy, String name) {
