+++ harrypotter/src/main/java/com/isep/harrypotter/model/others/Wand.java
@@ -1,6 +1,7 @@
-package com.isep.harrypotter;
+package com.isep.harrypotter.model.others;
 
 
+import com.isep.harrypotter.model.others.Core;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
