+++ harrypotter/src/main/java/com/isep/harrypotter/model/spells/AbstractSpell.java
@@ -1,14 +1,16 @@
-package com.isep.harrypotter.spells;
+package com.isep.harrypotter.model.spells;
 
 
-import com.isep.harrypotter.characters.Wizard;
+import com.isep.harrypotter.model.characters.Wizard;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 @AllArgsConstructor
 @Getter
 @Setter
+@NoArgsConstructor
 public abstract class AbstractSpell {
     private String name;
     private int energyCost;
@@ -16,6 +18,5 @@ public abstract class AbstractSpell {
 
     //TODO % de chance de réussite
 
-    public abstract String castSpell(String spellName, Wizard wizard);
 
 }
