+++ harrypotter/src/main/java/com/isep/harrypotter/view/InputParser.java
@@ -1,0 +1,10 @@
+package com.isep.harrypotter.view;
+
+import com.isep.harrypotter.model.spells.Spell;
+import com.isep.harrypotter.model.characters.Wizard;
+
+public interface InputParser {
+    int getInt(String messageWhenMismatch);
+    String getString(Wizard wizard);
+
+}
