+++ harrypotter/src/main/java/com/isep/utils/Tools.java
@@ -1,0 +1,4 @@
+package com.isep.utils;
+
+public class Tools implements ToolsManager {
+}
