+++ harrypotter/src/main/java/com/isep/utils/ToolsManager.java
@@ -1,0 +1,4 @@
+package com.isep.utils;
+
+public interface ToolsManager {
+}
