+++ harrypotter/src/main/java/module-info.java
@@ -6,8 +6,14 @@ module com.isep.harrypotter {
 
     opens com.isep.harrypotter to javafx.fxml;
     exports com.isep.harrypotter;
-    exports com.isep.harrypotter.characters;
-    opens com.isep.harrypotter.characters to javafx.fxml;
-    exports com.isep.harrypotter.spells;
-    opens com.isep.harrypotter.spells to javafx.fxml;
+    exports com.isep.harrypotter.model.characters;
+    opens com.isep.harrypotter.model.characters to javafx.fxml;
+    exports com.isep.harrypotter.model.spells;
+    opens com.isep.harrypotter.model.spells to javafx.fxml;
+    exports com.isep.harrypotter.model;
+    opens com.isep.harrypotter.model to javafx.fxml;
+    exports com.isep.harrypotter.controller;
+    opens com.isep.harrypotter.controller to javafx.fxml;
+    exports com.isep.harrypotter.model.others;
+    opens com.isep.harrypotter.model.others to javafx.fxml;
 }
